% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EMBinom.R
\name{binom.params.m.recur}
\alias{binom.params.m.recur}
\title{ECM: M-Step for Binomial expert.}
\usage{
binom.params.m.recur(
  binom.params.old,
  tl,
  yl,
  yu,
  tu,
  expert.ll,
  expert.tn,
  expert.tn.bar,
  z.e.obs,
  z.e.lat,
  k.e,
  penalty,
  hyper.params
)
}
\description{
ECM: M-Step for Binomial expert.
}
\seealso{
\code{\link{comp.kj.params.m.recur}}
}
\keyword{internal}
