% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SwitchPosEM.R
\name{comp.kj.params.m.recur}
\alias{comp.kj.params.m.recur}
\title{ECM: M-Step for updating parameters of positive part of component experts.}
\usage{
comp.kj.params.m.recur(
  comp.kj.dist,
  comp.kj.params.old,
  tl.k,
  yl.k,
  yu.k,
  tu.k,
  comp.kj.pos.expert.ll,
  comp.kj.pos.expert.tn,
  comp.kj.pos.expert.tn.bar,
  z.e.obs,
  z.e.lat,
  k.e,
  penalty,
  hyper.params.kj
)
}
\arguments{
\item{comp.kj.dist}{A string which indicates component distribution. See also \code{\link{pos.expert.loglik.calc}}.}

\item{comp.kj.params.old}{A vector of numerics: old parameter values for \code{comp.kj.dist}.}

\item{tl.k, yl.k, yu.k, tu.k}{Numerical vectors of length N. See also \code{\link{pos.expert.loglik.calc}}.}

\item{comp.kj.pos.expert.ll, comp.kj.pos.expert.tn, comp.kj.pos.expert.tn.bar}{See \code{\link{pos.expert.loglik.calc}}.}

\item{z.e.obs, z.e.lat, k.e}{A numerical vector returned by \code{\link{comp.zkz.e.recur}}.}

\item{penalty, hyper.params.kj}{See \code{\link{expert.loglik.pen.dim.comp}}.}
}
\value{
Updated parameter values.
}
\description{
ECM: M-Step for updating parameters of positive part of component experts.
}
\keyword{internal}
