% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EMzkz.R
\name{comp.zkz.e.recur}
\alias{comp.zkz.e.recur}
\title{ECM: E-Step for \code{z.e.obs}, \code{k.e} and \code{z.e.lat}.}
\usage{
comp.zkz.e.recur(gate.ll, expert.list, ll.list)
}
\arguments{
\item{gate.ll}{An object returned by \code{\link{gate.logit}}.}

\item{expert.list}{An object returned by \code{\link{expert.loglik.dim.comp}}.}

\item{ll.list}{An object returned by \code{\link{gate.expert.loglik}}.}
}
\value{
\code{z.e.obs},\code{k.e},\code{z.e.lat} Numerical vectors of length N.
}
\description{
ECM: E-Step for \code{z.e.obs}, \code{k.e} and \code{z.e.lat}.
}
\keyword{internal}
