% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CountParams.R
\name{count.alpha}
\alias{count.alpha}
\title{Count the number of logit regression coefficients.}
\usage{
count.alpha(alpha)
}
\arguments{
\item{alpha}{A g*P matrix of logit regression coefficients.}
}
\value{
The number of logit regression coefficients.
}
\description{
Count the number of logit regression coefficients.
}
\keyword{internal}
