% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CovInfContinuous.R
\name{covinf.continuous}
\alias{covinf.continuous}
\title{Investigate the influence of X (continuous covariate) on Y.}
\usage{
covinf.continuous(
  X,
  idx,
  eval.seq = seq(from = 1, to = 10, by = 1),
  response = c("Mean", "VAR990", "CTE990"),
  dim = 1,
  alpha,
  comp.dist,
  zero.prob,
  params.list
)
}
\arguments{
\item{X}{A matrix of covariates.}

\item{idx}{Column index of the continuous covariate of interest. There is no need to include the intercept.}

\item{eval.seq}{A numeric vector, containing values of the continuous covariate at which the  evaluation is done.}

\item{response}{A string vector indicating what metrics of Y to calculate. Currently supported calculation include:
\itemize{
   \item "Mean" : Mean of response
   \item "SD" : Variance of response
   \item "VAR990", "CTE990" : Value-at-Risk and Conditional Tail Expectation of response.
          The last three digits indicate the desired probability level, e.g. "990" = 99.0%.
}}

\item{dim}{The dimension of the response under consideration.}

\item{alpha}{A matrix of logit regression coefficients.}

\item{comp.dist}{A d*g matrix of strings, which specify the component distributions to fit.
The rows represent the dimensions of \code{Y}, while the columns represent the component distributions.}

\item{zero.prob}{A d*g matrix of numerics in (0,1), which specify the probability mass at zero for component distributions.}

\item{params.list}{A list of length d, where each element is a sublist of length g.
Each sublist contains one numeric vector, which is the parameter value for the corresponding \code{comp.dist}.}
}
\value{
A data frame containing all results.
}
\description{
Investigate the influence of X (continuous covariate) on Y.
}
\seealso{
\code{\link{LRMoE.fit}}.
}
