% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DemoData.R
\docType{data}
\name{DemoData}
\alias{DemoData}
\title{Demo data for \code{LRMoE}.}
\description{
\itemize{
  \item \code{X}: A matrix of covariates required by the LRMoE fitting function (complete).
  \item \code{Y}: A matrix of response required by the LRMoE fitting function (complete).
  \item \code{X.obs}: A matrix of covariates required by the LRMoE fitting function (after truncation and censoring).
  \item \code{Y.obs}: A matrix of response required by the LRMoE fitting function (after truncation and censoring).
}
}
\keyword{data}
