% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GammaCountNew.R
\name{dgammacount.new}
\alias{dgammacount.new}
\title{Modified \code{\link[rmutil]{GammaCount}} pmf for better numerical accuracy.}
\usage{
dgammacount.new(y, m, s, log = FALSE)
}
\arguments{
\item{y}{Vector of gamma count values.}

\item{m, s}{Paramaters of Gamma count distribution.}
}
\description{
Modified \code{\link[rmutil]{GammaCount}} pmf for better numerical accuracy.
}
\seealso{
\code{\link[rmutil]{GammaCount}}.
}
