% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExpertBurr.R
\name{expert.burr}
\alias{expert.burr}
\title{Expert Function: Burr.}
\usage{
expert.burr(tl, yl, yu, tu, g = 1, shape1.k, shape2.c, scale.lambda)
}
\arguments{
\item{tl}{A vector of length N: lower bounds of truncation.}

\item{yl}{A vector of length N: lower bounds of censoring.}

\item{yu}{A vector of length N: upper bounds of censoring.}

\item{tu}{A vector of length N: upper bounds of truncation.}

\item{g}{A numeric which is set to 1. See Note.}

\item{shape1.k}{A vector of length \code{g}: Burr shape1 parameters.}

\item{shape2.c}{A vector of length \code{g}: Burr shape2 parameters.}

\item{scale.lambda}{A vector of length \code{g}: Burr scale parameters.}
}
\value{
A list of matrices of expert loglikelihood for Burr.
}
\description{
Expert Function: Burr.
}
\section{Note}{

\code{g} is a legacy input. It is always set to 1, because this expert function is called within function \code{pos.expert.loglik.calc}.
}

\seealso{
\code{\link{pos.expert.loglik.calc}}, \code{\link[actuar]{Burr}}.
}
\keyword{internal}
