% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExpertGamma.R
\name{expert.gamma}
\alias{expert.gamma}
\title{Expert Function: Gamma.}
\usage{
expert.gamma(tl, yl, yu, tu, g = 1, m, theta)
}
\arguments{
\item{tl}{A vector of length N: lower bounds of truncation.}

\item{yl}{A vector of length N: lower bounds of censoring.}

\item{yu}{A vector of length N: upper bounds of censoring.}

\item{tu}{A vector of length N: upper bounds of truncation.}

\item{g}{A numeric which is set to 1. See Note.}

\item{m}{A vector of length \code{g}: Gamma shape parameters.}

\item{theta}{A vector of length \code{g}: Gamma scale parameters.}
}
\value{
A list of matrices of expert loglikelihood for Gamma.
}
\description{
Expert Function: Gamma.
}
\section{Note}{

\code{g} is a legacy input. It is always set to 1, because this expert function is called within function \code{pos.expert.loglik.calc}.
}

\seealso{
\code{\link{pos.expert.loglik.calc}}, \code{\link[stats]{GammaDist}}.
}
\keyword{internal}
