% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExpertGammaCount.R
\name{expert.gammacount}
\alias{expert.gammacount}
\title{Expert Function: Gamma Count.}
\usage{
expert.gammacount(tl, yl, yu, tu, g = 1, m, s)
}
\arguments{
\item{tl}{A vector of length N: lower bounds of truncation.}

\item{yl}{A vector of length N: lower bounds of censoring.}

\item{yu}{A vector of length N: upper bounds of censoring.}

\item{tu}{A vector of length N: upper bounds of truncation.}

\item{g}{A numeric which is set to 1. See Note.}

\item{m}{A vector of length \code{g}: Gamma Count shape parameters.}

\item{s}{A vector of length \code{g}: Gamma Count dispersion parameters.}
}
\value{
A list of matrices of expert loglikelihood for Gamma Count.
}
\description{
Expert Function: Gamma Count.
}
\section{Note}{

\code{g} is a legacy input. It is always set to 1, because this expert function is called within function \code{pos.expert.loglik.calc}.
}

\seealso{
\code{\link{pos.expert.loglik.calc}}, \code{\link[rmutil]{GammaCount}}.
}
\keyword{internal}
