% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExpertInverseGaussian.R
\name{expert.invgauss}
\alias{expert.invgauss}
\title{Expert Function: Inverse Gaussian.}
\usage{
expert.invgauss(tl, yl, yu, tu, g = 1, mean.mu, shape.lamba)
}
\arguments{
\item{tl}{A vector of length N: lower bounds of truncation.}

\item{yl}{A vector of length N: lower bounds of censoring.}

\item{yu}{A vector of length N: upper bounds of censoring.}

\item{tu}{A vector of length N: upper bounds of truncation.}

\item{g}{A numeric which is set to 1. See Note.}

\item{mean.mu}{A vector of length \code{g}: Inverse Gaussian mean parameters.}

\item{shape.lambda}{A vector of length \code{g}: Inverse Gaussian shape parameters.}
}
\value{
A list of matrices of expert loglikelihood for Inverse Gaussian.
}
\description{
Expert Function: Inverse Gaussian.
}
\section{Note}{

\code{g} is a legacy input. It is always set to 1, because this expert function is called within function \code{pos.expert.loglik.calc}.
}

\seealso{
\code{\link{pos.expert.loglik.calc}}, \code{\link[statmod]{invgauss}}.
}
\keyword{internal}
