% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExpertLogNormal.R
\name{expert.lognormal}
\alias{expert.lognormal}
\title{Expert Function: Log Normal.}
\usage{
expert.lognormal(tl, yl, yu, tu, g = 1, meanlog, sdlog)
}
\arguments{
\item{tl}{A vector of length N: lower bounds of truncation.}

\item{yl}{A vector of length N: lower bounds of censoring.}

\item{yu}{A vector of length N: upper bounds of censoring.}

\item{tu}{A vector of length N: upper bounds of truncation.}

\item{g}{A numeric which is set to 1. See Note.}

\item{meanlog}{A vector of length \code{g}: Lognormal mean parameters.}

\item{sdlog}{A vector of length \code{g}: Lognormal sd parameters.}
}
\value{
A list of matrices of expert loglikelihood for Log Normal.
}
\description{
Expert Function: Log Normal.
}
\section{Note}{

\code{g} is a legacy input. It is always set to 1, because this expert function is called within function \code{pos.expert.loglik.calc}.
}

\seealso{
\code{\link{pos.expert.loglik.calc}}, \code{\link[stats]{Lognormal}}.
}
\keyword{internal}
