% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExpertNegativeBinomial.R
\name{expert.nbinom}
\alias{expert.nbinom}
\title{Expert Function: Negative Binomial.}
\usage{
expert.nbinom(tl, yl, yu, tu, g = 1, size.n, prob.p)
}
\arguments{
\item{tl}{A vector of length N: lower bounds of truncation.}

\item{yl}{A vector of length N: lower bounds of censoring.}

\item{yu}{A vector of length N: upper bounds of censoring.}

\item{tu}{A vector of length N: upper bounds of truncation.}

\item{g}{A numeric which is set to 1. See Note.}

\item{size.n}{A vector of length \code{g}: Negative Binomial size.n parameters.}

\item{prob.p}{A vector of length \code{g}: Negative Binomial prob.p parameters.}
}
\value{
A list of matrices of expert loglikelihood for Negative Binomial.
}
\description{
Expert Function: Negative Binomial.
}
\section{Note}{

\code{g} is a legacy input. It is always set to 1, because this expert function is called within function \code{pos.expert.loglik.calc}.
}

\seealso{
\code{\link{pos.expert.loglik.calc}}, \code{\link[stats]{NegBinomial}}.
}
\keyword{internal}
