% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExpertPoisson.R
\name{expert.poisson}
\alias{expert.poisson}
\title{Expert Function: Poisson.}
\usage{
expert.poisson(tl, yl, yu, tu, g = 1, mean.theta)
}
\arguments{
\item{tl}{A vector of length N: lower bounds of truncation.}

\item{yl}{A vector of length N: lower bounds of censoring.}

\item{yu}{A vector of length N: upper bounds of censoring.}

\item{tu}{A vector of length N: upper bounds of truncation.}

\item{g}{A numeric which is set to 1. See Note.}

\item{mean.theta}{A vector of length \code{g}: Poisson mean parameters.}
}
\value{
A list of matrices of expert loglikelihood for Poisson.
}
\description{
Expert Function: Poisson.
}
\section{Note}{

\code{g} is a legacy input. It is always set to 1, because this expert function is called within function \code{pos.expert.loglik.calc}.
}

\seealso{
\code{\link{pos.expert.loglik.calc}}, \code{\link[stats]{Poisson}}.
}
\keyword{internal}
