% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExpertWeibull.R
\name{expert.weibull}
\alias{expert.weibull}
\title{Expert Function: Weibull.}
\usage{
expert.weibull(tl, yl, yu, tu, g = 1, shape.k, scale.lambda)
}
\arguments{
\item{tl}{A vector of length N: lower bounds of truncation.}

\item{yl}{A vector of length N: lower bounds of censoring.}

\item{yu}{A vector of length N: upper bounds of censoring.}

\item{tu}{A vector of length N: upper bounds of truncation.}

\item{g}{A numeric which is set to 1. See Note.}

\item{shape.k}{A vector of length \code{g}: Weibull shape parameters.}

\item{scale.lambda}{A vector of length \code{g}: Weibull scale parameters.}
}
\value{
A list of matrices of expert loglikelihood for Weibull.
}
\description{
Expert Function: Weibull.
}
\section{Note}{

\code{g} is a legacy input. It is always set to 1, because this expert function is called within function \code{pos.expert.loglik.calc}.
}

\seealso{
\code{\link{pos.expert.loglik.calc}}, \code{\link[stats]{Weibull}}.
}
\keyword{internal}
