% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EMGamma.R
\name{gamma.params.m.recur}
\alias{gamma.params.m.recur}
\title{ECM: M-Step for Gamma expert.}
\usage{
\method{gamma}{params.m.recur}(
  gamma.params.old,
  tl,
  yl,
  yu,
  tu,
  expert.ll,
  expert.tn,
  expert.tn.bar,
  z.e.obs,
  z.e.lat,
  k.e,
  penalty,
  hyper.params
)
}
\description{
ECM: M-Step for Gamma expert.
}
\seealso{
\code{\link{comp.kj.params.m.recur}}
}
\keyword{internal}
