% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EMGammaCount.R
\name{gammacount.params.m.recur}
\alias{gammacount.params.m.recur}
\title{ECM: M-Step for Gamma Count expert.}
\usage{
gammacount.params.m.recur(
  gammacount.params.old,
  tl,
  yl,
  yu,
  tu,
  expert.ll,
  expert.tn,
  expert.tn.bar,
  z.e.obs,
  z.e.lat,
  k.e,
  penalty,
  hyper.params
)
}
\description{
ECM: M-Step for Gamma Count expert.
}
\seealso{
\code{\link{comp.kj.params.m.recur}}
}
\keyword{internal}
