% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Loglik.R
\name{gate.expert.loglik}
\alias{gate.expert.loglik}
\title{Computes the entire loglikelihood, including gating function and experts.}
\usage{
gate.expert.loglik(
  alpha,
  gate.ll,
  expert.list,
  penalty = TRUE,
  hyper.alpha = NULL,
  hyper.params = NULL
)
}
\arguments{
\item{alpha}{A g*P matrix, where g is the number of components and P is the number of covariates.}

\item{gate.ll}{An object returned by function \code{\link{gate.logit}}.}

\item{expert.list}{An object returned by function \code{\link{expert.loglik.dim.comp}}.}

\item{penalty}{TRUE/FALSE, which indicates whether parameter penalty should be applied. Default (and recommended) is TRUE.}

\item{hyper.alpha}{A g*P matrix, which penalizes the magnitude of \code{alpha}.}

\item{hyper.params}{A list of length d. Each element is a sublist of length g.
Each element of a sublist is a vector of numerics, which penalizes expert parameters. See also \code{\link{expert.loglik.pen.dim.comp}}.}
}
\description{
Computes the entire loglikelihood, including gating function and experts.
}
\keyword{internal}
