% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GatingFunction.R
\name{gate.logit}
\alias{gate.logit}
\title{Computes the logit regression weights in log.}
\usage{
gate.logit(x, alpha)
}
\arguments{
\item{x}{An N * P covariate matrix, where N is sample size. The first column MUST be 1.}

\item{alpha}{A g * P matrix. Logit regression coefficients.}
}
\value{
The log of \code{tcrossprod(x,alpha)}, with each row normalized by \code{rowLogSumExps}
}
\description{
Computes the logit regression weights in log.
}
\keyword{internal}
