% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SwitchPenInit.R
\name{hyper.params.kj.initialize}
\alias{hyper.params.kj.initialize}
\title{Initialize penalty of parameters for expert functions by dimension and by component.}
\usage{
hyper.params.kj.initialize(ind.dist)
}
\arguments{
\item{ind.dist}{A string which indicates the expert function.
\itemize{
    \item \code{gamma}: Gamma
    \item \code{lnorm}: Log Normal
    \item \code{invgauss}: Inverse Gaussian
    \item \code{weibull}: Weibull
    \item \code{burr}: Burr
    \item \code{poisson}: Poisson
    \item \code{nbinom}: Negative Binomial
    \item \code{binom}: Binomial
    \item \code{gammacount}: Gamma Count
    \item \code{ZI-root}: Zero-inflated versions of the distributions above, e.g. \code{ZI-gamma}.
}}
}
\value{
A list of hyper parameters by dimension and by component.
}
\description{
Initialize penalty of parameters for expert functions by dimension and by component.
}
\seealso{
\code{\link{pos.expert.loglik.calc}}.
}
\keyword{internal}
