% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SwitchPosDist.R
\name{ind.cdf.y.pos}
\alias{ind.cdf.y.pos}
\title{A switch to calculate the cdf of a given distribution.}
\usage{
ind.cdf.y.pos(y, comp.dist, params)
}
\arguments{
\item{y}{A numeric vector at which the CDF is to be calculated.}

\item{comp.dist}{A string which indicates the expert function.
\itemize{
    \item \code{gamma}: Gamma
    \item \code{lnorm}: Log Normal
    \item \code{invgauss}: Inverse Gaussian
    \item \code{weibull}: Weibull
    \item \code{burr}: Burr
    \item \code{poisson}: Poisson
    \item \code{nbinom}: Negative Binomial
    \item \code{gammacount}: Gamma Count
    \item \code{ZI-root}: Zero-inflated versions of the distributions above, e.g. \code{ZI-gamma}.
}}

\item{params}{A vector of parameter values.
\itemize{
    \item \code{gamma}: \code{(m, theta)}
    \item \code{lnorm}: \code{(meanlog, sdlog)}
    \item \code{invgauss}: \code{(mean.mu, scale.lambda)}
    \item \code{weibull}: \code{(shape.k, scale.lambda)}
    \item \code{burr}: \code{(shape1.k, shape2.c, scale.lambda)}
    \item \code{poisson}: \code{(mean.theta)}
    \item \code{nbinom}: \code{(size.n, prob.p)}
    \item \code{gammacount}: \code{(m, s)}
}}
}
\description{
A switch to calculate the cdf of a given distribution.
}
\seealso{
\code{\link{LRMoE.fit}}, \code{\link{ind.dens.y.pos}}.
}
\keyword{internal}
