% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PredictLimitExcess.R
\name{ind.predict.limit}
\alias{ind.predict.limit}
\title{Predict the limited mean of y, given a fixed covariate vector X and a model.}
\usage{
ind.predict.limit(X, alpha, comp.dist, zero.prob, params.list, limit)
}
\arguments{
\item{X}{A matrix of covariates.}

\item{alpha}{A matrix of logit regression coefficients.}

\item{comp.dist}{A d*g matrix of strings, which specify the component distributions to fit.
The rows represent the dimensions of \code{Y}, while the columns represent the component distributions.}

\item{zero.prob}{A d*g matrix of numerics in (0,1), which specify the probability mass at zero for component distributions.}

\item{params.list}{A list of length d, where each element is a sublist of length g.
Each sublist contains one numeric vector, which is the parameter value for the corresponding \code{comp.dist}.}

\item{limit}{A vector of limit values by dimension.}
}
\value{
A vector of limited mean values by dimension.
}
\description{
Predict the limited mean of y, given a fixed covariate vector X and a model.
}
\seealso{
\code{\link{LRMoE.fit}}.
}
\keyword{internal}
