% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EMLogNormal.R
\name{lnorm.params.m.recur}
\alias{lnorm.params.m.recur}
\title{ECM: M-Step for LognOrmal expert.}
\usage{
lnorm.params.m.recur(
  lnorm.params.old,
  tl,
  yl,
  yu,
  tu,
  expert.ll,
  expert.tn,
  expert.tn.bar,
  z.e.obs,
  z.e.lat,
  k.e,
  penalty,
  hyper.params
)
}
\description{
ECM: M-Step for LognOrmal expert.
}
\seealso{
\code{\link{comp.kj.params.m.recur}}
}
\keyword{internal}
