% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LRMoELoglik.R
\name{LRMoE.loglik}
\alias{LRMoE.loglik}
\title{Computes the loglikelihood of LRMoE, given X, Y and a fitted model.}
\usage{
LRMoE.loglik(X, Y, model, penalty = TRUE, hyper.alpha, hyper.params)
}
\arguments{
\item{X}{A N*P matrix of covariates.}

\item{Y}{A N*d matrix of response.}

\item{model}{A list of parameters specifying an LRMoE model, including
\itemize{
     \item \code{alpha}: A g*P matrix, where g is the number of components and P is the number of covariates.
     \item \code{comp.dist}: A d*g matrix of strings, describing component distributions by dimension and by component.
     \item \code{zero.prob}: A d*g matrix of numbers between 0 and 1, describing zero probability masses by dimension and by component.
     \item \code{params.list}: A list of length d, where each element is a sublist of length g.
                   Each sublist contains one numeric vector, which is the initial parameter guess for the corresponding \code{comp.dist}.
       }}

\item{penalty}{TRUE/FALSE, which indicates whether parameter penalty should be applied. Default (and recommended) is TRUE.}

\item{hyper.alpha}{A numeric, which penalizes the magnitude of \code{alpha}.}

\item{hyper.params}{A list of length d. Each element is a sublist of length g.
Each element of a sublist is a vector of numerics, which penalizes expert parameters.}
}
\value{
Loglikelihood (with and without penalty), AIC and BIC.
}
\description{
Computes the loglikelihood of LRMoE, given X, Y and a fitted model.
}
