% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SwitchPosName.R
\name{name.y.pos}
\alias{name.y.pos}
\title{A function to format model output: add names for parameters}
\usage{
name.y.pos(comp.dist)
}
\description{
A function to format model output: add names for parameters
}
\seealso{
\code{\link{pos.expert.loglik.calc}}
}
\keyword{internal}
