% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GammaCountNew.R
\name{pgammacount.new}
\alias{pgammacount.new}
\title{Modified \code{\link[rmutil]{GammaCount}} cdf for better numerical accuracy.}
\usage{
pgammacount.new(q, m, s, log = FALSE)
}
\arguments{
\item{q}{Vector of quantiles.}

\item{m, s}{Paramaters of Gamma count distribution.}
}
\description{
Modified \code{\link[rmutil]{GammaCount}} cdf for better numerical accuracy.
}
\seealso{
\code{\link[rmutil]{GammaCount}}.
}
