% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotClassProbPost.R
\name{plot.ind.class.prob.posterior}
\alias{plot.ind.class.prob.posterior}
\title{Plots two stacked bar charts of latent class probabilities, which contrasts prior and posterior latent class probabilities, given a covariate vector.}
\usage{
\method{plot}{ind.class.prob.posterior}(
  X,
  Y,
  alpha,
  comp.dist,
  zero.prob,
  params.list,
  title = "Prediction of Latent Classes"
)
}
\arguments{
\item{X}{A matrix of covariates.}

\item{Y}{A matrix of observed responses for \code{X}.}

\item{alpha}{A matrix of logit regression coefficients.}

\item{comp.dist}{A d*g matrix of strings, which specify the component distributions to fit.
The rows represent the dimensions of \code{Y}, while the columns represent the component distributions.}

\item{zero.prob}{A d*g matrix of numerics in (0,1), which specify the probability mass at zero for component distributions.}

\item{params.list}{A list of length d, where each element is a sublist of length g.
Each sublist contains one numeric vector, which is the parameter value for the corresponding \code{comp.dist}.}
}
\value{
A \code{ggplot2} object.
}
\description{
Plots two stacked bar charts of latent class probabilities, which contrasts prior and posterior latent class probabilities, given a covariate vector.
}
\seealso{
\code{\link{LRMoE.fit}}, \code{\link{predict.class.prob.posterior}}
}
