% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotClassProb.R
\name{plot.ind.class.prob}
\alias{plot.ind.class.prob}
\title{Plots a stacked bar chart of latent class probabilities, given a covariate vector.}
\usage{
\method{plot}{ind.class.prob}(X, alpha, title = "Prediction of Latent Classes")
}
\arguments{
\item{X}{A vector of covariates for one policyholder.}

\item{alpha}{A matrix of logit regression coeffiecients.}

\item{title}{A text string for plot title.}
}
\value{
A \code{ggplot2} object.
}
\description{
Plots a stacked bar chart of latent class probabilities, given a covariate vector.
}
\seealso{
\code{\link{LRMoE.fit}}, \code{\link{predict.class.prob}}
}
