% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotIndDist.R
\name{plot.ind.fitted.dist}
\alias{plot.ind.fitted.dist}
\title{Plot the fitted density, given a vector of covariates and a model}
\usage{
\method{plot}{ind.fitted.dist}(
  X,
  alpha,
  comp.dist,
  zero.prob,
  params.list,
  plot.dim = 1,
  plot.lim = NULL
)
}
\arguments{
\item{X}{A vector of covariates.}

\item{alpha}{A g*P matrix of numerics, which contains initial guess of the logit regression coefficients.
The last row should all be zero, representing the default latent class.}

\item{comp.dist}{A d*g matrix of strings, which specify the component distributions to fit.
The rows represent the dimensions of \code{Y}, while the columns represent the component distributions.}

\item{zero.prob}{A d*g matrix of numerics in (0,1), which specify the probability mass at zero for component distributions.}

\item{params.list}{A list of length d, where each element is a sublist of length g.
Each sublist contains one numeric vector, which is the parameter value for the corresponding \code{comp.dist}.}

\item{plot.dim}{A numeric indicating which dimention of y to plot.}

\item{plot.lim}{Upper bound of y for plotting. Default is 50, if no value is provided.}
}
\value{
A \code{\link{ggplot2}} object.
}
\description{
Plot the fitted density, given a vector of covariates and a model
}
\seealso{
\code{\link{LRMoE.fit}}
}
