% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EMPoisson.R
\name{poisson.params.m.recur}
\alias{poisson.params.m.recur}
\title{ECM: M-Step for Poisson expert.}
\usage{
poisson.params.m.recur(
  poisson.params.old,
  tl,
  yl,
  yu,
  tu,
  expert.ll,
  expert.tn,
  expert.tn.bar,
  z.e.obs,
  z.e.lat,
  k.e,
  penalty,
  hyper.params
)
}
\description{
ECM: M-Step for Poisson expert.
}
\seealso{
\code{\link{comp.kj.params.m.recur}}
}
\keyword{internal}
