% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SwitchPosLoglik.R
\name{pos.expert.loglik.calc}
\alias{pos.expert.loglik.calc}
\title{A switch to calculate the loglikelihood of expert functions.}
\usage{
pos.expert.loglik.calc(ind.dist, tl, yl, yu, tu, params)
}
\arguments{
\item{ind.dist}{A string which indicates the expert function.
\itemize{
    \item \code{gamma}: Gamma
    \item \code{lnorm}: Log Normal
    \item \code{invgauss}: Inverse Gaussian
    \item \code{weibull}: Weibull
    \item \code{burr}: Burr
    \item \code{poisson}: Poisson
    \item \code{nbinom}: Negative Binomial
    \item \code{gammacount}: Gamma Count
    \item \code{ZI-root}: Zero-inflated versions of the distributions above, e.g. \code{ZI-gamma}.
}}

\item{tl}{A vector of length N: lower bounds of truncation.}

\item{yl}{A vector of length N: lower bounds of censoring.}

\item{yu}{A vector of length N: upper bounds of censoring.}

\item{tu}{A vector of length N: upper bounds of truncation.}

\item{params}{A vector of parameters for \code{ind.dist}.}
}
\value{
A list of matrices of expert loglikelihood for the chosen distribution \code{ind.dist}.
\itemize{
    \item \code{expert.ind.dist.ll}: An N*1 matrix of loglikelihood: density for exact observation; probability for censored observation.
    \item \code{expert.ind.dist.tn}: An N*1 matrix of loglikelihood within truncation limits.
    \item \code{expert.ind.dist.tn.bar}: An N*1 matrix of loglikelihood outside truncation limits, which is \code{log(1-exp(expert.ind.dist.tn))}.
}
}
\description{
A switch to calculate the loglikelihood of expert functions.
}
\seealso{
\code{\link{expert.gamma}}, \code{\link{expert.lognormal}}, \code{\link{expert.invgauss}}, \code{\link{expert.weibull}}, \code{\link{expert.burr}},
         \code{\link{expert.poisson}}, \code{\link{expert.nbinom}}, \code{\link{expert.gammacount}}
}
\keyword{internal}
