% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PredictClassProbPost.R
\name{predict.class.prob.posterior}
\alias{predict.class.prob.posterior}
\title{Predict the posterior latent class probabilities, given a fixed covariate matrix X, Y and a model.}
\usage{
\method{predict}{class.prob.posterior}(X, Y, alpha, comp.dist, zero.prob, params.list)
}
\arguments{
\item{X}{A matrix of covariates.}

\item{Y}{A matrix of observed responses for \code{X}.}

\item{alpha}{A matrix of logit regression coefficients.}

\item{comp.dist}{A d*g matrix of strings, which specify the component distributions to fit.
The rows represent the dimensions of \code{Y}, while the columns represent the component distributions.}

\item{zero.prob}{A d*g matrix of numerics in (0,1), which specify the probability mass at zero for component distributions.}

\item{params.init}{A list of length d, where each element is a sublist of length g.
Each sublist contains one numeric vector, which is the parameter value for the corresponding \code{comp.dist}.}
}
\value{
A matrix of latent class probabilities by observation and by component.
}
\description{
Predict the posterior latent class probabilities, given a fixed covariate matrix X, Y and a model.
}
\seealso{
\code{\link{LRMoE.fit}}.
}
