% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PredictClassProb.R
\name{predict.class.prob}
\alias{predict.class.prob}
\title{Predict the latent class probabilities, given a fixed covariate matrix X and a model.}
\usage{
\method{predict}{class.prob}(X, alpha)
}
\arguments{
\item{X}{A matrix of covariates.}

\item{alpha}{A matrix of logit regression coefficients.}
}
\value{
A matrix of latent class probabilities by observation and by component.
}
\description{
Predict the latent class probabilities, given a fixed covariate matrix X and a model.
}
\seealso{
\code{\link{LRMoE.fit}}.
}
