% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PredictClass.R
\name{predict.class}
\alias{predict.class}
\title{Predict the most likely latent class, given a fixed covariate matrix X and a model.}
\usage{
\method{predict}{class}(X, alpha)
}
\arguments{
\item{X}{A matrix of covariates.}

\item{alpha}{A matrix of logit regression coefficients.}
}
\value{
A vector of the most likely latent class by observation.
}
\description{
Predict the most likely latent class, given a fixed covariate matrix X and a model.
}
\seealso{
\code{\link{LRMoE.fit}}.
}
