% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PredictLimitExcess.R
\name{predict.excess}
\alias{predict.excess}
\title{Predict the excess mean of y, given a fixed covariate matrix X, a model and a vector of limits by dimension.}
\usage{
\method{predict}{excess}(X, alpha, comp.dist, zero.prob, params.list, limit)
}
\arguments{
\item{X}{A matrix of covariates.}

\item{alpha}{A matrix of logit regression coefficients.}

\item{comp.dist}{A d*g matrix of strings, which specify the component distributions to fit.
The rows represent the dimensions of \code{Y}, while the columns represent the component distributions.}

\item{zero.prob}{A d*g matrix of numerics in (0,1), which specify the probability mass at zero for component distributions.}

\item{params.list}{A list of length d, where each element is a sublist of length g.
Each sublist contains one numeric vector, which is the parameter value for the corresponding \code{comp.dist}.}

\item{limit}{A vector of limit to apply to each dimension of y.}
}
\value{
A matrix of excess mean values by observation and by dimension.
        This is equan to \code{predict.mean - predict.limit}.
}
\description{
Predict the excess mean of y, given a fixed covariate matrix X, a model and a vector of limits by dimension.
}
\seealso{
\code{\link{LRMoE.fit}}, \code{\link{predict.limit}}.
}
