% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SwitchPosSim.R
\name{sim.y.pos}
\alias{sim.y.pos}
\title{A function to simulate positive part of observations y}
\usage{
sim.y.pos(sample.size.n, comp.dist, params)
}
\arguments{
\item{sample.size.n}{The size of y to simulate.}
}
\description{
A function to simulate positive part of observations y
}
\seealso{
\code{\link{pos.expert.loglik.calc}}
}
\keyword{internal}
