% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EMWeibull.R
\name{weibull.params.m.recur}
\alias{weibull.params.m.recur}
\title{ECM: M-Step for Weibull expert.}
\usage{
weibull.params.m.recur(
  weibull.params.old,
  tl,
  yl,
  yu,
  tu,
  expert.ll,
  expert.tn,
  expert.tn.bar,
  z.e.obs,
  z.e.lat,
  k.e,
  penalty,
  hyper.params
)
}
\description{
ECM: M-Step for Weibull expert.
}
\seealso{
\code{\link{comp.kj.params.m.recur}}
}
\keyword{internal}
