% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EMzzerolat.R
\name{z.zero.e.lat.recur}
\alias{z.zero.e.lat.recur}
\title{ECM: E-Step for \code{z.zero.e.lat}.}
\usage{
z.zero.e.lat.recur(
  tl.k,
  comp.kj.zero.inflation,
  comp.kj.zero.prob.old,
  comp.kj.pos.expert.tn.bar
)
}
\arguments{
\item{tl.k}{A vector of length N: lower bounds of truncation.}

\item{comp.kj.zero.inflation}{TRUE/FALSE. See returned value \code{zero.inflation} in \code{\link{expert.loglik.dim.comp}}.}

\item{comp.kj.zero.prob.old}{A numeric between 0 and 1. See returned value \code{zero.prob} in \code{\link{expert.loglik.dim.comp}}.}

\item{comp.kj.pos.expert.tn.bar}{See returned value \code{pos.expert.tn.bar} in \code{\link{expert.loglik.dim.comp}}}
}
\value{
\code{z.zero.e.lat} Numerical vectors of length N.
}
\description{
ECM: E-Step for \code{z.zero.e.lat}.
}
\keyword{internal}
