% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EMzzeroobs.R
\name{z.zero.e.obs.recur}
\alias{z.zero.e.obs.recur}
\title{ECM: E-Step for \code{z.zero.e.obs}.}
\usage{
z.zero.e.obs.recur(
  yl.k,
  comp.kj.zero.inflation,
  comp.kj.zero.prob.old,
  comp.kj.pos.expert.ll
)
}
\arguments{
\item{yl.k}{A vector of length N: lower bounds of censoring.}

\item{comp.kj.zero.inflation}{TRUE/FALSE. See returned value \code{zero.inflation} in \code{\link{expert.loglik.dim.comp}}.}

\item{comp.kj.zero.prob.old}{A numeric between 0 and 1. See returned value \code{zero.prob} in \code{\link{expert.loglik.dim.comp}}.}

\item{comp.kj.pos.expert.ll}{See returned value \code{pos.expert.ll} in \code{\link{expert.loglik.dim.comp}}}
}
\value{
\code{z.zero.e.obs} Numerical vectors of length N.
}
\description{
ECM: E-Step for \code{z.zero.e.obs}.
}
\keyword{internal}
