% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EMzeroprob.R
\name{zero.prob.m.recur}
\alias{zero.prob.m.recur}
\title{ECM: M-Step for \code{zero.prob}.}
\usage{
zero.prob.m.recur(z.zero.e.obs, z.pos.e.obs, z.zero.e.lat, z.pos.e.lat, k.e)
}
\arguments{
\item{z.zero.e.obs}{An object returned by \code{\link{z.zero.e.obs.recur}}.}

\item{z.pos.e.obs}{An object returned by \code{\link{comp.zkz.e.recur}}.}

\item{z.zero.e.lat}{An object returned by \code{\link{z.zero.e.lat.recur}}.}

\item{z.pos.e.lat}{An object returned by \code{\link{comp.zkz.e.recur}}.}

\item{k.e}{An object returned by \code{\link{comp.zkz.e.recur}}.}
}
\value{
\code{zero.prob} Updated zero.prob.
}
\description{
ECM: M-Step for \code{zero.prob}.
}
\keyword{internal}
